added royale api to .env and begun implemented more of the api
noticed that with one of the api calls to royale api that I could see daily clan stats. The great thing about this is that it contains a date which means I can take that date and check to see if a player that joined on that date joined on the same date as a past war which could be used to remove the bug that shows new players as having missed 3 wars in a row.